requirements.txt
pandas
numpy
scikit-learn
joblib
PyYAML
flask
pytest
flake8

